Move test_data to sample_data
