+++ metadata-mapper/metadata_mapper/sample_data/nuxeo_harvests.py
